requirements.txt
google-adk>=0.4.0
google-cloud-bigquery>=3.11.0
fastmcp>=0.4.0
streamlit>=1.32.0
plotly>=5.20.0
pandas>=2.0.0
google-genai>=0.5.0
